condensed camera warp and vision warp
