+++ src/vision_warp.cpp
@@ -14,52 +14,34 @@
 #include <opencv2/calib3d.hpp>
 
 #include "vision_warp.h"
+#include <libterraclear/src/filetools.hpp>
 
 using namespace cv;
 using namespace std;
 
 namespace  terraclear
 {   
-    vision_warp::vision_warp() 
+    vision_warp::vision_warp(std::string exe_filepath) 
     {
         _sw.start();
+        
+        // Get path to camera calibration file
+        std::string basepath = terraclear::filetools::get_base_path(exe_filepath);
+        std::string cam_file = basepath + "/../../../camera.xml";
+        // Read in camera intrinsics 
+        cv::FileStorage fs(cam_file, FileStorage::READ);
+        fs["camera_matrix"] >> cameraMatrix;
+        fs["distortion_coefficients"] >> distCoeffs;
     }
 
     vision_warp::~vision_warp() 
     {
     }
-    
-    void vision_warp::init_transform()
+        
+    void vision_warp::update_frame(const cv::Mat img_src)
     {
-        //Source & Dest Points
-        std::vector<cv::Point2f> points_src;
-        points_src.push_back(_source_points.top_left);        
-        points_src.push_back(_source_points.top_right);        
-        points_src.push_back(_source_points.bottom_right);        
-        points_src.push_back(_source_points.bottom_left); 
-        
-        //calculate target points based on target size..
-        roi_transform target_points;
-        target_points.top_left.x = 0;
-        target_points.top_left.y = 0;
-        target_points.top_right.x = _target_size.width;
-        target_points.top_right.y = 0;
-        target_points.bottom_right.x = _target_size.width;
-        target_points.bottom_right.y = _target_size.height;
-        target_points.bottom_left.x = 0;
-        target_points.bottom_left.y = _target_size.height;
-        
-        std::vector<cv::Point2f> points_dst;
-        points_dst.push_back(target_points.top_left);        
-        points_dst.push_back(target_points.top_right);        
-        points_dst.push_back(target_points.bottom_right);        
-        points_dst.push_back(target_points.bottom_left); 
-
-        //compute transformation matrix.
-        _transform_matrix =  cv::findHomography(points_src, points_dst); 
-
+        _img = img_src;
     }
-    
     cv::Mat vision_warp::get_transfor_matrix()
     {
         return _transform_matrix;
@@ -216,69 +198,76 @@ namespace  terraclear
         return dst_corners;
     }
     
-    cv::Mat vision_warp::get_chessboard_transform(cv::Mat img_src, cv::Size board_sz)
-    {
-        cv::Mat gray_image;
-        cv::cvtColor(img_src, gray_image, cv::COLOR_BGR2GRAY);
-        
-        /* Look for chessboard corners */
-        std::vector<Point2d> corners;
-        
-        bool found = cv::findChessboardCorners(gray_image, board_sz, corners,
-                    CALIB_CB_ADAPTIVE_THRESH + CALIB_CB_NORMALIZE_IMAGE + CALIB_CB_FAST_CHECK);
-        
-        /* If we find the chessboard */
-        if (found)
-        {
-            //Get four corners of the board
-            std::vector<Point2d> org_corners;
+    bool vision_warp::findChessBoard(const cv::Size board_sz) {
+        // Find chess board corners 
+        return findChessboardCorners(_img, board_sz, corners);
+    }
     
-            org_corners.push_back(cv::Point((int)(std::floor(corners[board_sz.area() - 1].x)),(int)(std::floor(corners[board_sz.area() - 1].y))));
-            org_corners.push_back(cv::Point((int)(std::floor(corners[board_sz.area() - board_sz.width].x)),(int)(std::floor(corners[board_sz.area() - board_sz.width].y))));
-            org_corners.push_back(cv::Point((int)(std::floor(corners.begin()->x)),(int)(std::floor(corners.begin()->y))));
-            org_corners.push_back(cv::Point((int)(std::floor(corners[board_sz.width-1].x)),(int)(std::floor(corners[board_sz.width-1].y))));
-            //org_corners.push_back(cv::Point((img_src.cols-700)/2,img_src.cols));
-            //org_corners.push_back(cv::Point((img_src.cols+700)/2,img_src.cols));
-            for (int i = 0; i < corners.size(); i++)
-            {
-                corners[i].x = std::floor(corners[i].x);
-                corners[i].y = std::floor(corners[i].y);
-            }
-            struct myclass {
-            bool operator() (cv::Point pt1, cv::Point pt2) { return (pt1.y < pt2.y);}
-            } myobject;
-            std::sort(corners.begin(), corners.end(), myobject);
+    void vision_warp::calcChessboardCorners(cv::Size boardSize, float squareSize, terraclear::Pattern patternType = terraclear::Pattern::CHESSBOARD)
+    {
+        objectPoints.resize(0);
 
-            int miny = (corners.begin())->y;
-
-            struct myclass2 {
-            bool operator() (cv::Point pt1, cv::Point pt2) { return (pt1.x < pt2.x);}
-            } myobject1;
-            std::sort(corners.begin(), corners.end(), myobject1);
-
-            int minx = (corners.begin())->x;
-            int maxx = (--corners.end())->x;
-
-            // Need to get corners of warped board for finding homography
-            std::vector<Point2d> corr_corners;
-            // Corners of the rectified chessboard
-            corr_corners.push_back(cv::Point(minx,miny));
-            corr_corners.push_back(cv::Point(maxx,miny));
-            // Keep original bottom corners of the board so that the bottom of the image isnt warped out of frame
-            corr_corners.push_back(org_corners[2]);
-            corr_corners.push_back(org_corners[3]);
-            
-            //corr_corners.push_back(cv::Point((img_src.cols-700)/2,img_src.cols));
-            //corr_corners.push_back(cv::Point((img_src.cols+700)/2,img_src.cols));
-            
-            _transform_matrix = cv::findHomography(org_corners, corr_corners, cv::RANSAC);
-
-            return _transform_matrix;   
-        }
-        else 
+        switch (patternType)
         {
-            std::string o = "o";
+        case CHESSBOARD:
+        case CIRCLES_GRID:
+            //! [compute-chessboard-object-points]
+            for( int i = 0; i < boardSize.height; i++ )
+                for( int j = 0; j < boardSize.width; j++ )
+                    //To try to center the chessboard frame, we substract the image size
+                    objectPoints.push_back(cv::Point3f(float((j-boardSize.width/2)*squareSize),
+                                              float((i-boardSize.height/2)*squareSize), 0));
+            //! [compute-chessboard-object-points]
+            break;
+
+        case ASYMMETRIC_CIRCLES_GRID:
+            for( int i = 0; i < boardSize.height; i++ )
+                for( int j = 0; j < boardSize.width; j++ )
+                    objectPoints.push_back(cv::Point3f(float((2*j + i % 2)*squareSize),
+                                              float(i*squareSize), 0));
+            break;
+
+        default:
+            CV_Error(Error::StsBadArg, "Unknown pattern type\n");
         }
+    }
+
+    void vision_warp::computeC2MC1(const Mat &R1, const Mat &tvec1, const Mat &R2, const Mat &tvec2,
+                      Mat &R_1to2, Mat &tvec_1to2)
+    {
+        //c2Mc1 = c2Mo * oMc1 = c2Mo * c1Mo.inv()
+        R_1to2 = R2 * R1.t();
+        tvec_1to2 = R2 * (-R1.t()*tvec1) + tvec2;
+    }
+    
+    cv::Mat vision_warp::init_transform()
+    {
+        cv::Mat rvec, tvec;
+        cv::solvePnP(objectPoints, corners, cameraMatrix, distCoeffs, rvec, tvec);
+
+        cv::Mat R_desired = (Mat_<double>(3,3) <<
+                        -1, 0, 0,
+                        0, -1, 0,
+                        0, 0, 1);
+        cv::Mat R;
+        cv::Rodrigues(rvec, R);
+        cv::Mat normal = (Mat_<double>(3,1) << 0, 0, 1);
+        cv::Mat normal1 = R*normal;
+        cv::Mat origin(3, 1, CV_64F, Scalar(0));
+        cv::Mat origin1 = R*origin + tvec;
+        
+        double d_inv1 = 1.0 / normal1.dot(origin1);
+        
+        cv::Mat R_1to2, tvec_1to2;
+        cv::Mat tvec_desired = tvec.clone();
+
+        vision_warp::computeC2MC1(R, tvec, R_desired, tvec_desired, R_1to2, tvec_1to2);
+       
+        _H = R_1to2 + d_inv1 * tvec_1to2*normal1.t();
+
+        _H = cameraMatrix * _H * cameraMatrix.inv();
+        
+        return  _H/_H.at<double>(2,2);
             
     }
